Add files via upload
Arduino ide code that displays the voltage and analog signal of a 5v power source, with those corresponding values will turn on and off a group of LED's, the LED will be decided by conditional statements. A button and its corresponding status is used to decide whether the circuit is running or not through the button pin state and a Boolean variable. If the button's state is Low the circuit will run until the button is pressed and turn back on when it is pressed again.

Used 4 LED's, a Push Button, Potentiometer, Arduino Uno R3, Breadboard.